Created component for meals planning. Fix some issues in HttpClient extensions.

+++ src/FoodPlannerBlazor.Domain/Entities/PlannedMeal/Outgoing/CreatePlannedMeal.cs
@@ -2,5 +2,5 @@
 
 namespace FoodPlannerBlazor.Domain.Entities.PlannedMeal.Outgoing
 {
-    public record CreatePlannedMeal(byte OrdinalNumber, DateTime ScheduledFor, int MealId);
+    public record CreatePlannedMeal(int OrdinalNumber, DateTime ScheduledFor, int MealId);
 }
