+++ src/FoodPlannerBlazor.Domain/Entities/PlannedMeal/PlannedMeal.cs
@@ -5,7 +5,7 @@ namespace FoodPlannerBlazor.Domain.Entities.PlannedMeal
 {
     public class PlannedMeal : BaseEntity
     {
-        public byte OrdinalNumber { get; set; }
+        public int OrdinalNumber { get; set; }
         public DateTime ScheduledFor { get; set; }
         public Meal.Meal Meal { get; set; } = new();
     }
