+++ src/FoodPlannerBlazor.Domain/Entities/PlannedMeal/PlannedMealForGrouping.cs
@@ -4,7 +4,7 @@ namespace FoodPlannerBlazor.Domain.Entities.PlannedMeal
 {
     public class PlannedMealForGrouping : BaseEntity
     {
-        public byte OrdinalNumber { get; set; }
+        public int OrdinalNumber { get; set; }
         public Meal.Meal Meal { get; set; }
     }
 }
